+++ src/Contracts/EmizorApiContract.php
@@ -1,6 +1,7 @@
 <?php
 namespace Emizor\SDK\Contracts;
 
+use Closure;
 use Emizor\SDK\DTO\RegisterDTO as RegisterCredentialsDTO;
 use Emizor\SDK\Enums\ParametricType;
 use Emizor\SDK\Exceptions\EmizorApiConnectionTimeoutException;
@@ -20,14 +21,35 @@ interface EmizorApiContract {
 
 
     /**
+     * List parametrics types
+     * @return array
+     */
+    public function listParametricsTypes():array;
+
+    /**
      * Sync parametrics by type, according to account will decide if specific or global
      * @param ParametricType $type
      */
-    public function sync(array $parametrics):void;
+    public function syncParametrics(array $parametrics):void;
 
     /**
      * Get parametrics by type, according to account will decide if specific or global
-     * @param ParametricType $type
+     * @param string $type
      */
-    public function getParametric(ParametricType $type):array;
+    public function getParametric($type):array;
+
+    /**
+     * Define defaults for this account, using a closure-style builder.
+     *
+     * @param Closure $callback
+     * @return self
+     */
+    public function setDefaults(Closure $callback): self;
+
+    /**
+     * Get the current defaults applied for this account.
+     *
+     * @return array
+     */
+    public function getDefaults(): array;
 }