+++ src/Contracts/HttpClientInterface.php
@@ -6,18 +6,21 @@ interface HttpClientInterface
 {
     /**
      * Set the host to be used in requests.
+     *
+     * @param string $host
      */
     public function withBaseUri(string $host): static;
 
     /**
      * Set the token to be used in requests.
+     *
+     * @param string $token
      */
     public function withToken(string $token): static;
 
     /**
      * Make a GET request to a dynamic host.
      *
-     * @param string $host
      * @param string $uri
      * @param array $options
      */
@@ -26,7 +29,6 @@ interface HttpClientInterface
     /**
      * Make a POST request to a dynamic host.
      *
-     * @param string $host
      * @param string $uri
      * @param array $data
      * @param array $options
@@ -36,7 +38,6 @@ interface HttpClientInterface
     /**
      * Make a PUT request to a dynamic host.
      *
-     * @param string $host
      * @param string $uri
      * @param array $data
      * @param array $options
@@ -46,7 +47,6 @@ interface HttpClientInterface
     /**
      * Make a DELETE request to a dynamic host.
      *
-     * @param string $host
      * @param string $uri
      * @param array $options
      */