+++ src/Contracts/ParametricContract.php
@@ -6,15 +6,26 @@ use Emizor\SDK\Enums\ParametricType;
 
 interface ParametricContract
 {
+
 /**
-* Sincroniza desde API y guarda en DB.
+* Set Host for getting parametrics
+*/
+public function setHost( string $host ): static;
+
+/**
+* Set Token for getting parametrics
+*/
+public function setToken( string $token ): static;
+
+/**
+* Sync from API and save in DB
 */
 public function sync( $type, string $accountId): void;
 
 /**
 * Obtiene paramétricas ya sincronizadas desde DB.
 *
-* @return ParametricDTO[]
+* @return array
 */
 public function get( $type, string $accountId): array;
 }