+++ src/Enums/InvoiceType.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Emizor\SDK\Enums;
+
+enum InvoiceType: string
+{
+    case COMPRA_VENTA = 'compra-venta';
+
+}