+++ src/Facade/EmizorSdk.php
@@ -2,10 +2,12 @@
 
 namespace Emizor\SDK\Facade;
 
+use Emizor\SDK\Enums\ParametricType;
 use Illuminate\Support\Facades\Facade;
 
 class EmizorSdk extends Facade
 {
+
     protected static function getFacadeAccessor()
     {
         return 'emizorsdk';
@@ -18,6 +20,11 @@ class EmizorSdk extends Facade
     public static function withAccount(string $accountId)
     {
         return app('emizorsdk', ['accountId' => $accountId]);
+    }
+
+    public static function PARAMETRICS_TYPES()
+    {
+        return app('emizorsdk')->listParametricsTypes();
     }