+++ changelog.md
@@ -5,6 +5,10 @@
 - Added generate token
 - Added sync parametric, global and specific
 - Added list parametric
+- Added set defaults
+- Added service list type parametrics
 
 ## Changed
 - Changed use token implementation in services
+- Changed parametric service instance in emizorapi implementation
+- Changed type parametrics, validation